reorganize and protect patches
